+++ src/rex.instrument/NEWS.rst
@@ -21,7 +21,7 @@ REX.INSTRUMENT Change Log
   Instrument definitions.
 - The ``instrument-retrieve`` rex command now accepts a ``format`` option to
   indicate that you want JSON or YAML returned.
-- Added a ``definition_yaml`` propert to the InstrumentVersion class to get or
+- Added a ``definition_yaml`` property to the InstrumentVersion class to get or
   set the Instrument definition using YAML.